Move preprocessing methods to util
